rename suggestions menu to suggestion list
